04_Production_RAG/requirements.txt
langchain==0.3.15
langchain-core==0.3.31
langchain-community==0.3.15
langchain-openai==0.3.1
langchain-qdrant==0.2.0
qdrant-client==1.13.2
tiktoken
pymupdf==1.25.2
python-dotenv
